promote CCGraphViewBuilder to be the only GraphViewBuilder
* delete the non CC compatible GraphViewBuilder
* rename the CCGraphViewBuilder to drop the CC prefix
